requirements.txt
pymupdf==1.24.9
pytesseract==0.3.13
pillow==10.4.0
tqdm==4.66.5
regex==2024.7.24
wordfreq==3.1.1